+++ __docs__/wonder-blocks-tabs/navigation-tab-item-variants.stories.tsx
@@ -102,6 +102,7 @@ const generateRows = (rtl: boolean = false) => {
             props: {
                 children: (
                     <IconButton
+                        kind="tertiary"
                         href="#link"
                         aria-label="Ice cream"
                         icon={IconMappings.iceCream}