+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 5.1.6
+
+### Patch Changes
+
+- 4846e9c: Update icon buttons to use the new mapping (from `primary` to `tertiary`)
+- Updated dependencies [3dc5dac]
+- Updated dependencies [4846e9c]
+- Updated dependencies [4846e9c]
+    - @khanacademy/wonder-blocks-tokens@5.2.0
+    - @khanacademy/wonder-blocks-icon-button@8.0.0
+    - @khanacademy/wonder-blocks-form@7.1.5
+
+## 5.1.5
+
+### Patch Changes
+
+- Updated dependencies [61f7837]
+- Updated dependencies [86e1901]
+- Updated dependencies [3cacbe7]
+- Updated dependencies [4887c59]
+    - @khanacademy/wonder-blocks-icon-button@7.0.0
+
 ## 5.1.4
 
 ### Patch Changes