+++ packages/wonder-blocks-tabs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-tabs
 
+## 0.2.2
+
+### Patch Changes
+
+- 4846e9c: Fix pseudo-states to account for aria-disabled
+- Updated dependencies [3dc5dac]
+    - @khanacademy/wonder-blocks-tokens@5.2.0
+
+## 0.2.1
+
+### Patch Changes
+
+- 86e1901: Override border on different states (focus-visible, hover, press) to account for the `IconButton` change from `outline` to `border` + `boxShadow`.
+
 ## 0.2.0
 
 ### Minor Changes